classProject/requirements.txt
azure-functions
Flask
pandas
fredapi
pyodbc
requests
matplotlib
numpy